+++ lib/nanites/option.rb
@@ -1,66 +1,94 @@
 # frozen_string_literal: true
 
 module Nanites
-  # Generic value class which either holds an instance of [Some] or [None]
-  # :reek:MissingSafeMethod { exclude: [ value! ] }
-  class Option
-    attr_reader :value
-
-    # Create a new Option value
-    def initialize(_ = nil); end
-
-    # Is a [None] value?
-    def none?
-      is_a? None
-    end
-
-    # Is a [Some] value?
-    def some?
-      is_a? Some
-    end
-
-    # Insecure getter for value, overwritten in descendants
-    def value!
-      raise NotImplementedError
-    end
-
+  # Generic value which either holds a value ([Some]) or not ([None])
+  module Option
     class << self
-      # Helper returning an instance of [Some]
+      # Create a container for an unknown value
+      # @param [value] value option value
+      # @return [Some] if not nil value was given
+      # @return [None] if nil value was given
+      def option(value = nil)
+        value ? some(value) : none
+      end
+
+      # Create a not nil value container
+      # @param [Object] value An arbitrary value
       def some(value)
         Some.new value
       end
 
-      # Helper returning an instance of [None]
+      # Create a nil value container
       def none
         None.new
       end
     end
+
+    # @abstract
+    # Is a Some value?
+    def some?; end
+
+    # @abstract
+    # Is a None value?
+    def none?; end
+
+    # @abstract
+    # Safe getter for value
+    # @returns [Object]
+    def value; end
+
+    # @abstract
+    # Unsafe getter for value
+    # @returns [Object]
+    # @raise [Nanites::Errors::ValueError]
+    def value!; end
   end
 
-  # Generic value representing a not nil value to be used inside an Option instance
+  # Generic class representing a not nil value
   # :reek:MissingSafeMethod { exclude: [ value! ] }
-  class Some < Option
-    # Create a new value container
-    # @param [Object] value an arbitrary object
+  class Some
+    include Option
+
+    attr_reader :value
+
     def initialize(value)
-      super
       @value = value
     end
 
-    # Getter for encapsulated value object
-    # @raise [Nanites::Errors::ValueError] if value is nil
+    # @see Option#value!
     def value!
       raise Nanites::Errors::ValueError, 'Value is nil' unless @value
 
-      value
+      @value
+    end
+
+    # @see Option#some?
+    def some?
+      true
+    end
+
+    # @see Option#none?
+    def none?
+      false
     end
   end
 
-  # Generic value representing a nil or empty value to be used inside an Option instance
+  # Generic class representing a nil value
   # :reek:MissingSafeMethod { exclude: [ value! ] }
-  class None < Option
-    # Convenience method raising a ValueError if called
-    # @raise [Nanites::Errors::ValueError]
+  class None
+    include Option
+
+    # @see Option#some?
+    def some?
+      false
+    end
+
+    # @see Option#none?
+    def none?
+      true
+    end
+
+    # @see Option#value!
     def value!
       raise Nanites::Errors::ValueError, "No values available in a #{self.class.name} object"
     end