+++ spec/other/none_spec.rb
@@ -8,7 +8,7 @@ RSpec.describe Nanites::None do
   context 'when initializing' do
     it 'can be initialized w/ value' do
       expect do
-        described_class.new 'foo'
+        described_class.new
       end.to_not raise_error
     end
 
@@ -21,7 +21,7 @@ RSpec.describe Nanites::None do
 
   context 'when calling value' do
     it 'raises ValueError w/ value' do
-      none = described_class.new 'foo'
+      none = described_class.new
 
       expect do
         none.value!
@@ -29,7 +29,7 @@ RSpec.describe Nanites::None do
     end
 
     it 'returns nil value if value is present' do
-      none = described_class.new 'foo'
+      none = described_class.new
 
       expect(none.value).to be_nil
     end
@@ -38,6 +38,16 @@ RSpec.describe Nanites::None do
       expect do
         described_class.new.value!
       end.to raise_error Nanites::Errors::ValueError
+    end
+  end
+
+  context 'when determining value type' do
+    it 'returns true for some?' do
+      expect(described_class.new.some?).to be_falsey
+    end
+
+    it 'returns false for none?' do
+      expect(described_class.new.none?).to be_truthy
     end
   end
 end