+++ lib/nanites/errors.rb
@@ -5,5 +5,8 @@ module Nanites
   module Errors
     # Is raised, when a None option is tried to be unwrapped
     class ValueError < StandardError; end
+
+    # Is raised, when a None option is tried to be unwrapped
+    class NoCommandError < StandardError; end
   end
 end