+++ spec/other/result_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-RSpec.describe Nanites::Actions::Result do
+RSpec.describe Nanites::Commands::Result do
   it_behaves_like :ClassLoader
 
   let :none do