+++ spec/support/spec_command.rb
@@ -2,7 +2,7 @@
 
 # Sample command used for the specs
 # Assumes no payload is given
-class SpecCommand < Nanites::Actions::Command
+class SpecCommand < Nanites::Commands::Command
   def execute(**params)
     if params.empty?
       error params, 'No parameters defined'
@@ -13,15 +13,27 @@ class SpecCommand < Nanites::Actions::Command
   end
 end
 
-class PayloadSpecCommand < Nanites::Actions::Command
+class PayloadSpecCommand < Nanites::Commands::Command
   def execute(**params)
     success @payload, "Got a #{@payload.class.name} object as payload"
     error nil, 'Got nil as payload' unless @payload
   end
 end
 
-class AlwaysErrorCommand < Nanites::Actions::Command
+class AlwaysErrorCommand < Nanites::Commands::Command
   def execute(**params)
     raise StandardError, 'Will always be raised'
+  end
+end
+
+class NoneCommand  < Nanites::Commands::Command
+  def execute(**params)
+    success nil, 'Payload will always be nil'
+  end
+end
+
+class SomeCommand  < Nanites::Commands::Command
+  def execute(**params)
+    success 'foo', 'Payload will always be something'
   end
 end