+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.1.1 - 2021-12-24
 
-minor improvements, complemented README
+refactored building Some/None values, complemented README
 
 ## 0.1.0 - 2021-12-23