+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Releases
 
+## 0.1.2 - 2021-12-24
+
+refactored module names, refactored compound, added more convenient compounds, complemented README
+
 ## 0.1.1 - 2021-12-24
 
 refactored building Some/None values, complemented README